🍱 Remove the Communities directly from workers
